Create unimplemented contains5 function
Created the contains5 function in a cpp file and have it throw a string
when it is called.

I am doing this so as to ensure that the compiler is able to compile the
program, since I do not intend to use a return value.

I also plan to use this function in assert macros with inputs and outputs
based on the sample results given in the challenge post so that I can
check if the function works.