Add assertions to main method
To make it absolutely clear that the type is std::string, I declared
variables to hold the input and expected output.
Also, I do not intend to have implicit conversions get in the way, so by
using the type directly, I do not have to worry about any side effects.