Create test for numbers 21 to 29
Created a test to check the output of the convert() method
for numbers 21 to 29.

The main goal is to check if a number like 25 or 36 will give us the
ten's name, followed by the one's name.

It's too much to test from 21 all the way to 99. Just testing a
subset should be good enough.