Create tasks.json
As I'm using Visual Studio Code to write this C++ program, I have to
create tasks to run when building the program.

By right the tasks.json file should be ignored locally if this project
is to be worked on by multiple people with different IDEs, but I decided
to just leave it here for my future reference and for anyone who wants
to take a look at it.